% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{temp_conversion}
\alias{temp_conversion}
\title{Temperature conversion}
\usage{
temp_conversion(temp_f)
}
\arguments{
\item{temp_f}{A numerical vector in F}
}
\value{
A numerical vector in C
}
\description{
It converts the temperature from F to C.
}
\examples{
temp_conversion(1)
}
